requirements.txt
fastapi[standard]
uvicorn
python-dotenv
livekit-api
assemblyai
requests
google-cloud-translate==2.0.1